+++ edge-util/src/Microsoft.Azure.Devices.Edge.Util/metrics/IEdgeMetrics.cs
@@ -1,0 +1,8 @@
+// Copyright (c) Microsoft. All rights reserved.
+namespace Microsoft.Azure.Devices.Edge.Util.Metrics
+{
+    public interface IEdgeMetrics
+    {
+        void InitPrometheusMetrics(int port);
+    }
+}
