+++ edge-util/src/Microsoft.Azure.Devices.Edge.Util/metrics/IMetricsCounter.cs
@@ -1,0 +1,13 @@
+// Copyright (c) Microsoft. All rights reserved.
+namespace Microsoft.Azure.Devices.Edge.Util.Metrics
+{
+    using System.Collections.Generic;
+
+    public interface IMetricsCounter
+    {
+        void Increment(long amount);
+        void Decrement(long amount);
+        void Increment(long amount, Dictionary<string, string> tags);
+        void Decrement(long amount, Dictionary<string, string> tags);
+    }
+}
