+++ edge-util/src/Microsoft.Azure.Devices.Edge.Util/metrics/nullMetrics/NullDisposable.cs
@@ -1,0 +1,16 @@
+// Copyright (c) Microsoft. All rights reserved.
+namespace Microsoft.Azure.Devices.Edge.Util.Metrics.NullMetrics
+{
+    using System;
+
+    public class NullDisposable : IDisposable
+    {
+        NullDisposable() { }
+
+        public static IDisposable Instance = new NullDisposable();
+
+        public void Dispose()
+        {
+        }
+    }
+}
