+++ edge-util/src/Microsoft.Azure.Devices.Edge.Util/metrics/nullMetrics/NullMetricsCounter.cs
@@ -1,0 +1,21 @@
+// Copyright (c) Microsoft. All rights reserved.
+namespace Microsoft.Azure.Devices.Edge.Util.Metrics.NullMetrics
+{
+    using System.Collections.Generic;
+
+    public class NullMetricsCounter : IMetricsCounter
+    {
+        public void Increment(long amount) { }
+
+        public void Decrement(long amount)
+        {
+        }
+
+        public void Increment(long amount, Dictionary<string, string> tags)
+        {
+        }
+
+        public void Decrement(long amount, Dictionary<string, string> tags)
+        { }
+    }
+}
