+++ edge-util/src/Microsoft.Azure.Devices.Edge.Util/metrics/nullMetrics/NullMetricsHistogram.cs
@@ -1,0 +1,16 @@
+// Copyright (c) Microsoft. All rights reserved.
+namespace Microsoft.Azure.Devices.Edge.Util.Metrics.NullMetrics
+{
+    using System.Collections.Generic;
+
+    public class NullMetricsHistogram : IMetricsHistogram
+    {
+        public void Update(long value)
+        {
+        }
+
+        public void Update(long value, Dictionary<string, string> tags)
+        {
+        }
+    }
+}
